+++ include/roq/api.hpp
@@ -84,7 +84,7 @@
 #include "roq/modify_order.hpp"
 #include "roq/order_ack.hpp"
 #include "roq/order_update.hpp"
-#include "roq/parameter_update.hpp"
+#include "roq/parameters_update.hpp"
 #include "roq/position_update.hpp"
 #include "roq/rate_limit_trigger.hpp"
 #include "roq/reference_data.hpp"