+++ include/roq/client/event_log_reader.hpp
@@ -50,7 +50,7 @@ class ROQ_PUBLIC EventLogReader {
     virtual void operator()(Event<CustomMetricsUpdate> const &) = 0;
 
     // parameters
-    virtual void operator()(Event<ParameterUpdate> const &) = 0;
+    virtual void operator()(Event<ParametersUpdate> const &) = 0;
   };
 
   virtual ~EventLogReader() {}