+++ include/roq/parameters_update.hpp
@@ -9,6 +9,7 @@
 #include <fmt/format.h>
 
 #include <span>
+#include <string_view>
 
 #include "roq/event.hpp"
 #include "roq/name.hpp"
@@ -19,52 +20,55 @@
 namespace roq {
 
 //! Update relating to parameters
-struct ROQ_PUBLIC ParameterUpdate final {
+struct ROQ_PUBLIC ParametersUpdate final {
   std::span<Parameter> parameters;  //!< List of parameters
   UpdateType update_type = {};      //!< Update type
+  std::string_view user;            //!< User name (optional, only relevant for drop-copy)
 };
 
 template <>
-inline constexpr std::string_view get_name<ParameterUpdate>() {
+inline constexpr std::string_view get_name<ParametersUpdate>() {
   using namespace std::literals;
-  return "parameter_update"sv;
+  return "parameters_update"sv;
 }
 
 }  // namespace roq
 
 template <>
-struct fmt::formatter<roq::ParameterUpdate> {
+struct fmt::formatter<roq::ParametersUpdate> {
   template <typename Context>
   constexpr auto parse(Context &context) {
     return std::begin(context);
   }
   template <typename Context>
-  auto format(roq::ParameterUpdate const &value, Context &context) const {
+  auto format(roq::ParametersUpdate const &value, Context &context) const {
     using namespace std::literals;
     return fmt::format_to(
         context.out(),
         R"({{)"
         R"(parameters=[{}], )"
-        R"(update_type={})"
+        R"(update_type={}, )"
+        R"(user="{}")"
         R"(}})"sv,
         fmt::join(value.parameters, ", "sv),
-        value.update_type);
+        value.update_type,
+        value.user);
   }
 };
 
 template <>
-struct fmt::formatter<roq::Event<roq::ParameterUpdate>> {
+struct fmt::formatter<roq::Event<roq::ParametersUpdate>> {
   template <typename Context>
   constexpr auto parse(Context &context) {
     return std::begin(context);
   }
   template <typename Context>
-  auto format(roq::Event<roq::ParameterUpdate> const &event, Context &context) const {
+  auto format(roq::Event<roq::ParametersUpdate> const &event, Context &context) const {
     using namespace std::literals;
     return fmt::format_to(
         context.out(),
         R"({{)"
-        R"(parameter_update={}, )"
+        R"(parameters_update={}, )"
         R"(message_info={})"
         R"(}})"sv,
         event.value,
@@ -73,18 +77,18 @@ struct fmt::formatter<roq::Event<roq::ParameterUpdate>> {
 };
 
 template <>
-struct fmt::formatter<roq::Trace<roq::ParameterUpdate>> {
+struct fmt::formatter<roq::Trace<roq::ParametersUpdate>> {
   template <typename Context>
   constexpr auto parse(Context &context) {
     return std::begin(context);
   }
   template <typename Context>
-  auto format(roq::Trace<roq::ParameterUpdate> const &event, Context &context) const {
+  auto format(roq::Trace<roq::ParametersUpdate> const &event, Context &context) const {
     using namespace std::literals;
     return fmt::format_to(
         context.out(),
         R"({{)"
-        R"(parameter_update={}, )"
+        R"(parameters_update={}, )"
         R"(trace_info={})"
         R"(}})"sv,
         event.value,